+++ stencil-workspace/src/components/modus-tooltip/modus-tooltip.spec.ts
@@ -5,14 +5,14 @@ describe('modus-tooltip', () => {
   it('renders', async () => {
     const { root } = await newSpecPage({
       components: [ModusTooltip],
-      html: '<modus-tooltip></modus-tooltip>',
+      html: '<modus-tooltip text="message"></modus-tooltip>',
     });
     expect(root).toEqualHtml(`
-      <modus-tooltip>
+      <modus-tooltip text="message">
         <mock:shadow-root>
           <div class="modus-tooltip top">
             <slot></slot>
-            <div class="text" role="tooltip"></div>
+            <div class="text" role="tooltip">message</div>
           </div>
         </mock:shadow-root>
       </modus-tooltip>